requirements.txt
deepgram-sdk
python-dotenv
aiofiles
tqdm
PyYAML
pydub
watchdog
moviepy
backoff